% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_features_reduction.R
\name{map_acoustic_nonlinear}
\alias{map_acoustic_nonlinear}
\title{Non-Linear Dimensionality Reduction Wrapper (t-SNE & UMAP)}
\usage{
map_acoustic_nonlinear(features, method = "umap", dims = 2, perplexity = 30)
}
\arguments{
\item{features}{A data frame or matrix of scaled numeric features.}

\item{method}{Character string: "tsne" or "umap" (default: "umap").}

\item{dims}{Integer. Number of dimensions for the output projection (default: 2).}

\item{perplexity}{Numeric. The perplexity parameter, specifically used if method is "tsne" (default: 30).}
}
\value{
A data frame with exactly two columns: "Dim1" and "Dim2".
}
\description{
A safe wrapper function to project high-dimensional acoustic data into 2D space for visualization using manifold learning.
}
\details{
This function safely wraps the \code{Rtsne} and \code{uwot} packages. It includes fail-safes such as setting \code{n_threads = 1} for UMAP to prevent C++ deadlock crashes on Windows machines, and automatically capping t-SNE perplexity to prevent it from exceeding the dataset row limits.
}
\examples{

data(bird_train)
# Scale data before UMAP projection
scaled_feats <- scale(as.matrix(bird_train[, 1:15]))

# Create a 2D UMAP embedding
umap_2d <- map_acoustic_nonlinear(scaled_feats, method = "umap", dims = 2)
plot(umap_2d$Dim1, umap_2d$Dim2)

}
